Rename time_strategies to strategies
