+++ client/verta/tests/strategies.py
